refactor: extract shared GOT-patching primitives into libdd-got-hook
Move the ELF parsing, dl_iterate_phdr iteration, PageProtGuard,
gnu_hash_symbol_count, gnu_hash_lookup, lookup_symbol, and related
utilities out of libdd-profiling-heap-gotter into a new libdd-got-hook
crate.

libdd-profiling-heap-gotter now depends on libdd-got-hook and keeps
only the SymbolOverrides multi-symbol registry and per-library
dedup/rescan logic. This is a pure code move — no behavioral changes.
