+++ libdd-profiling-heap-gotter/Cargo.toml
@@ -25,6 +25,7 @@ test-support = []
 
 [target.'cfg(target_os = "linux")'.dependencies]
 libc.workspace = true
+libdd-got-hook = { version = "1.0.0", path = "../libdd-got-hook" }
 libdd-profiling-heap-sampler = { version = "1.0.0", path = "../libdd-profiling-heap-sampler" }
 
 [target.'cfg(target_os = "linux")'.dev-dependencies]
